move fadeinslideout to a motion dir, update toaster
