+++ packages/gamut/src/Motion/FadeInSlideOut.tsx
